requirements.txt
yfinance
pandas
numpy
python-dotenv
gspread
oauth2client
streamlit
fastapi
uvicorn
schedule
requests
